Update 01_merge_multiple_datasets_for_classification to be more similar to real-life use case
