+++ datumaro/components/visualizer.py
@@ -2,6 +2,7 @@
 #
 # SPDX-License-Identifier: MIT
 import math
+import random
 import warnings
 from collections import defaultdict
 from typing import Iterable, List, Optional, Tuple, Union
@@ -122,6 +123,8 @@ class Visualizer:
         assert 0.0 <= alpha <= 1.0, "alpha should be in [0, 1]."
         self.alpha = alpha
 
+        self._items = [item for item in self.dataset]
+
     @property
     def draw_only_image(self):
         """
@@ -192,31 +195,74 @@ class Visualizer:
 
         return sorted(annotations, key=_sort_key)
 
+    def get_random_items(self, n_samples: int) -> List[DatasetItem]:
+        """Get random samples from the dataset"""
+        if n_samples >= len(self.dataset):
+            raise ValueError(
+                f"n_samples={n_samples} should be less than the dataset size ({len(self.dataset)})."
+            )
+
+        return random.choices(self._items, k=n_samples)
+
     def vis_gallery(
         self,
-        ids: List[Union[str, DatasetItem]],
-        subset: Optional[Union[str, List[str]]] = None,
+        ids: Optional[List[str]] = None,
+        subsets: Optional[Union[str, List[str]]] = None,
+        items: Optional[List[DatasetItem]] = None,
         grid_size: Tuple[Optional[int], Optional[int]] = (None, None),
     ) -> Figure:
+        self._check_inputs(ids, subsets, items)
+
+        if isinstance(subsets, str):
+            subsets = [subsets] * len(ids)  # expand it to have len(ids)
+
+        def _parse_inputs(
+            ids: Optional[List[str]],
+            subsets: Optional[List[str]],
+            items: Optional[List[DatasetItem]],
+        ) -> Tuple[List[str], List[str]]:
+            if ids is not None and subsets is not None:
+                return ids, subsets
+            elif items is not None:
+                return [item.id for item in items], [item.subset for item in items]
+            raise ValueError(
+                f"ids={ids}, subsets={subsets}, and items={items} is an invalid input."
+            )
+
+        ids, subsets = _parse_inputs(ids, subsets, items)
+
         nrows, ncols = _infer_grid_size(len(ids), grid_size)
         fig, axs = plt.subplots(nrows, ncols, figsize=self.figsize)
 
-        if isinstance(subset, list):
-            assert len(ids) == len(
-                subset
-            ), "If subset is a list, it should have the same length as ids."
+        assert len(ids) == len(
+            subsets
+        ), "If subset is a list, it should have the same length as ids."
 
-        for i, (dataset_id, ax) in enumerate(zip(ids, axs.flatten())):
-            if isinstance(subset, List):
-                self.vis_one_sample(dataset_id, subset[i], ax)
-            else:
-                self.vis_one_sample(dataset_id, subset, ax)
+        for item_id, subset, ax in zip(ids, subsets, axs.flatten()):
+            self.vis_one_sample(item_id=item_id, subset=subset, ax=ax)
 
         return fig
 
+    @staticmethod
+    def _check_inputs(
+        item_id: Optional[str] = None,
+        subset: Optional[str] = None,
+        item: Optional[DatasetItem] = None,
+    ) -> None:
+        if (item_id is None and subset is None) and item is None:
+            raise ValueError("(item_id and subset) or item should be not None.")
+        if item_id is not None and subset is not None and item is not None:
+            raise ValueError("Both (item_id and subset) and item cannot be provided.")
+
     def vis_one_sample(
-        self, id: Union[str, DatasetItem], subset: Optional[str] = None, ax: Optional[Axes] = None
+        self,
+        item_id: Optional[str] = None,
+        subset: Optional[str] = None,
+        item: Optional[DatasetItem] = None,
+        ax: Optional[Axes] = None,
     ) -> Figure:
+        self._check_inputs(item_id, subset, item)
+
         if ax is None:
             fig = plt.figure(figsize=self.figsize)
             ax = plt.gca()
@@ -224,8 +270,23 @@ class Visualizer:
             fig = ax.get_figure()
             plt.sca(ax)
 
-        item: DatasetItem = self.dataset.get(id, subset)
-        assert item is not None, f"Cannot find id={id}, subset={subset}"
+        def _parse_inputs(
+            item_id: Optional[str], subset: Optional[str], item: Optional[DatasetItem]
+        ) -> Tuple[str, str]:
+            if item_id is not None and subset is not None:
+                return item_id, subset
+            elif item is not None:
+                item_id = item.id
+                subset = item.subset
+                return item_id, subset
+            raise ValueError(
+                f"item_id={item_id}, subset={subset}, and item={item} is an invalid input."
+            )
+
+        item_id, subset = _parse_inputs(item_id, subset, item)
+        item: DatasetItem = self.dataset.get(item_id, subset)
+
+        assert item is not None, f"Cannot find id={item_id}, subset={subset}"
         assert (
             item is not Image
         ), f"Media type should be Image, Current media type={type(item.media)}"
@@ -235,7 +296,7 @@ class Visualizer:
         ax.imshow(img)
 
         width = ax.transAxes.transform_point((1, 0))[0] - ax.transAxes.transform_point((0, 0))[0]
-        text = ax.set_title(f"ID: {id}, Subset: {subset}", loc="center", wrap=True)
+        text = ax.set_title(f"ID: {item_id}, Subset: {subset}", loc="center", wrap=True)
         text.__get_wrapped_text = text._get_wrapped_text
 
         def _get_wrapped_text():
